cv/requirements.txt
numpy
opencv-python
matplotlib
pandas
Pillow
rasterio
